requirements.txt
google-genai>=0.1.0
pinecone-client>=4.0.0
python-dotenv>=1.0.0
